server/long_speech_to_text/requirements.txt
google-cloud-storage
google-cloud-speech
flask